Create ExpandablePane widget
Remove button element and its event handlers and move it to its own UI
class.

Search preview is disbaled for the time being, some vestiges of how the
functionality will be implemented in a separate UI class can be seen in
ExpandablePane. Also, code for generating the labels is still in init
and will also be moved to the new class.

Change-Id: I1a972326862f383f7e70c7be00c97474a96c78e4