Enforce namespace defaults in model
Always keep the main namespace in model, even when it's removed.
Always keep the file namespace in model when the filetype option is
selected.
To avoid infinite loops, don't trigger update event when the actual
namespaces don't change. This could have happened when the Article
namespace is removed by the user.

NB: Removing the article namespace as the last namespace in the
NamespaceFilter makes iot appear to be empty, however the UI for the
enformced namespaces (e.g. removing the close buttons) still has to be
done.

Bug: T167802
Change-Id: I5db8d33c2e97e49f1451182ca1642e27c0f041b5