Avoid poluting the browsers address bar with irrelevant options
I noticed that, when I do nothing but just clicking the "Search" button,
the URL that shows up in my browsers address bar is "poluted" with all
options, even the ones I did not even used.

…&advancedSearch-current={
    "options":{
        "filew":[">",""],
        "fileh":[">",""],
        "intitle":"",
        "or":[],
        "filetype":"",
        "phrase":[],
        "not":""
    },
    "namespaces":["6"]
}

The code change in this patch is untested and might be incomplete. It
just tries to reduce the amount of not needed information. Please feel
welcome to continue working on this patch.

Bug: T188509
Change-Id: I82c13101c555e0b8bc4ced08be955d275e148e92