ui.ArbitraryWordInput: Use util.arrayEquals() instead of a duplicate
Also fix dependencies on the 'ext.advancedSearch.util' module in
extension.json. It was added as a dependency in a few modules that
don't use its only function, and was missing in the only module that
actually uses it. (None of this matters anyway because the modules
are never loaded separately.)

Change-Id: I6f8b5e547fe75f896ddbf91b0f0ad69908ef289c