Mirror state of namespace selection in indicator
Added a workaround fix to show initial status on load correctly. There
seems to be some event that will set the menu open on load even though
it is not.

This could not be fixed by emitting the toggle event manually in the 
same function call after constructing the interface so I did the hack
with the async call.

Bug: T183346
Change-Id: Idc193f07048292f8f1cf40fff57367415f2d788c