Improve SearchPage toggling
The search_submit test was failing randomly, this PR uses a more
specific selector to toggle (open/close) the advanced search
fields, because the previously selected element (the whole pane,
including the search fields) was too big of a click target
that might have not been viable all the time. Now the expansion is
toogled by the little arrow icon.

Bug: T186703
Change-Id: Id39a01cf543b3ca58bb9288013f9ca502190993d