Leave original search field in place
Instead of replacing the contents of the search field with some
"original search" value and switching the HTML around on submit,
remove the advanced search part of the field when the field is loaded.

This assumes that the advanced search query is always the right-most
part of the search value.

Moved query compiling and search query modification to its own tested
class.

Bug: T188493
Change-Id: I4d11b815f949bd7353daddefdf4dc4a8fe2fe0d3