Respect ns… parameters when no namespace state is available
This patch suggests a few minor optimizations:

* In getNamespacesFromUrl, instead of building an array and filtering
it afterwards, do the filtering in the first loop. This should be a little
faster as it avoids constructing a second array.

* Keep the namespaces in object form, and check if an object key exists
with the "in" keyword. That should be much faster than scanning an array
with indexOf.

* I turned the two places around that set the namespace selection from
either the ns… URL parameters, or from the AdvancedSearch state. Now the
ns… parameters are processed first. They are only overruled if the
AdvancedSearch state contains a namespace selection. I find this edge case
relevant. Before the code actively ignored the ns… parameters, even if no
other namespace selection was there.

Bug: T182254
Change-Id: I3c671b25085a353cf205bd0cfc9480228cb65470