Replace all parseInt() with Number()
We don't really need parsing capabilities here. All we really need is a
type conversion from strings that are guaranteed to be integer-like to
integers. Number() does exactly this.

This patch also contains some performance optimizations. E.g. when
iterating the users options the regex matching is skipped in case the
value is false anyway.

This greatly reduces the complexity of the custom sort() function in
SearchModel.sortNamespacesByNumber. This exact use case is even listed as
an example on
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/sort

The codebase is made more consistent:
* Arrays that contain namespace numbers, but no labels, are named
  "namespaceIds" now.
* Objects mapping namespace IDs to labels are iterated via Object.keys().

I can easily split individual changes to separate patches if this makes
reviewing this easier. Please tell me.

Change-Id: Ie1491a64147d9612c5504596cb0c2754117ba793