Reduce visual jumps while the JavaScript is loaded
This greatly reduces the visual jumps that appear while the JavaScript
is loaded. Well, at least a significant amount of the jumping that
happens for several different reasons.

Intentionally not part of this patch: For some reason the entire block
of "advanced parameters" is expanded for the blink of an eye. This
creates a super-short but very visible distraction, because the affected
area is so large.

Bug: T189028
Change-Id: I228dc5549061e1735fdf5b0ad221f8f19fd9a96a