Unselect namespace dropdown items before disabling them
This patch does not fully fix T190156, but makes it less confusing. I
still believe we are dealing with an upstream bug that causes the gray
highlight to appear to late. E.g. it does not appear when I type "m",
but appears when I type "me".

This patch removes the blue selection from list items that are going to
be disabled anyway. This greatly reduces the amount of different
highlighting colors in the dropdown list. Disabled elements get gray now,
and not a confusing mixture of blue and gray.

Bug: T190156
Change-Id: I6534976b2de5282335ef0e1804e415cbba094269