Restore broken namespace selection and remember feature
This got removed in I17f54b8 (T181638) and broke a lot of features,
including the "remember selection for future searches" feature itself.
Steps to reproduce:

Try to remove a namespace from your user options by removing it and
checking "remember". It won't work. The namespace will always come back.
This is because the old, duplicate checkboxes never get removed from the
DOM.

Bug: T181638
Change-Id: Id32fa03bec62733cf76709cd11d533ac5edbd3b1