Remove preset labels from tests
These labels are not even part of what is actually tested. The labels
are also not really needed to make the tests readable. The presets are
named anyway via the object keys "justatest: …", "all: …", and so on. In
some cases the label does not even match this object key.

I feel the test setup becomes more condensed and focused with this being
removed, that's why I'm suggesting it here.

Change-Id: I3238ffa8b58d6edeaf8f27955c30e049015ed3a8