Update slectors in search.page.js
We realized the click() is literally done on a random position on the
element. Sometimes there is a namespace pill that would receive the click.
Even if the widget would be fine with this, the webdriver reports this as
an error.

I feel more comfortable with using a class name in the selector, and not
an element name. Some OOUI widgets might have a second, hidden <input>
somewhere. I feel the class name selector is a little more on the safe
side.

Change-Id: Iab9edcc98b46363b158c282d42ed9abb3d5156a4