Fix "Default" checkbox not appearing because of rigorous filtering
The new "Default" checkbox relies on the users settings. These
setting can contain stuff that looks like a valid namespace setting,
but contains a number that is not a valid namespace. This is not a bug,
but possible for quite a lot of reasons. (And it actually happened on
my local machine, that's how I found this issue.) The previous code
ditched the preset entirely if this happend. The "Default" checkbox
just disappeared entirely. This might have been reasonable before for
the other presets, but can not be done any more for this new "Default"
preset.

Bug: T188515
Change-Id: I59548f866a0ce407e98f3af9bb29b7bee7d2e4b6