Split AdvancedOptionsBuilder
Split configuration and widget building into two classes and uncouple
the configuration from the SearchModel.

Clean up dependencies - init does not need the UI classes for search
fields, those are best placed in AdvancedOptionsConfig where they are
used.

This refactoring is in preparation for adding field defaults to the
options which will be used in SearchModel.

Bug: T190681
Change-Id: I3e57bd2225cfe3e76d1c6df73360b9ba2ecfb67b