Fix FOUC by explicitely specifying dependency order on .less files
The problem with the original specification was that there was no order
specified for the two files "ext.advancedSearch.init.js" and
"ext.advancedSearch.less". Which gets loaded first was unclear. But the
two files actually depend on each other. The .js creates elements and
adds them to the visible page. The styles for these elements must
already be present at this point. Otherwise a flash of unstyled content
(FOUC) happens for a few milliseconds until the styles are loaded.

This happened all the time on my dev environment. It seems it's harder
to reproduce on production. I figured it's worth being explicit about
this loading order no matter what. So if this does not change anything
for you, it should still be merged. :-)

Bug: T189028
Change-Id: I74fc004aba345a3a61e6d61a14627c532d5a3351