Delay all initialization until advanced options panel is expanded
This is a suprisingly trivial patch that does one thing: The big pane
with all the advanced search options that are hidden by default is not
created when the page loads, but later when the user clicks the arrow
down button for the first time. This speeds up the original loading time.
It's very relevant for all use cases where the user is *not* going to use
any advanced option. The entire panel is not needed then.

Note I had to remove the possibility to have an ExpandablePane expanded
by default. This is not used anyway.

This is only a first step to cut the initialization time down. The next
step will be to split the ResourceLoader modules and delay all that are
not required to initialize the collapsed UI. It turns out this is hard
because mw.libs.advancedSearch.AdvancedOptionsConfig contains all
initialization code, even if this is not called when the collapsed UI is
initialized. How to split this properly?

Bug: T200083

Change-Id: I5e4da55dd7c02cab14efc96efcfdf23c7a607671