Timeless: Remove inconsistent resets on <code> elements
The Timeless skin intentionally sets the colors of all <code> elements
to different values. The two lines I'm removing in this patch got
introduced via Id8ef926, but introduced an inconsistency for Timeless
users: All <code> elements use the same intended colors for Timeless
users, except on Special:Search. This is confusing and should not be
done. We need to either leave the colors as intended by Timeless, or
remove them entirely from the Timeless skin.

Bug: T192303
Change-Id: I1e802f5955da2308d97db07d43c49698d54be181