Do not set CSS values via (expensive) JavaScript
I'm not exactly sure why this was done in Id8ef926, but it apparently
is not necessary.

I'm also reorganizing the Timeless overrides a little bit to make it
much more visible what the purpose of these overrides is: all this is
only relevant in the info box popups.

Note this patch here intentionally clashes with I7d58d4a, where I
remove some of these overrides. Both are valid approaches: Either we
intentionally ignore all design decisions Timeless made and reset
everything in our popups back to the original Wikimedia design (this is
what this patch here does). Or we go with Timeless and accept what it
does to <h6>, <dt>, and <code>.

Both is inconsistent: Either AdvancedSearch looks different in different
skins, or identical elements (e.g. <code>) look different on different
pages within the same skin.

Bug: T192303
Change-Id: I143395adbfe82a08a205d4daccd993c813999148