Make randomly failing browser tests less flaky
The "inserts content in icon popups" keeps failing on a more or less
constant basis for the past days. The current theory is: There is some
other popup on the page (e.g. from the Echo extension), that messes up
the order of popup elements in the array returned by this selector. By
limiting the selector to popups *inside* of AdvancedSearch this should
not be possible any more.

An other test that started failing is one that clicks the "Search"
button, but then forgot to wait for the page to load.

Change-Id: If86038504df9f214688ad14dbaa866776b8baff9