Remove empty lines from beginning/end of function scopes
This is a pure style question, and really meant as a question: Can the
team that owns this code decide on this? I noticed that the current code
(even within one file) does this very inconsistently.

Note this is only about empty lines on function/method levels, not on
the top class level, where several style guides encourage to use an
empty line.

The MediaWiki CodeSniffer we are using on PHP code contains a sniff that
disallows newlines at the begining/end of function scopes. I suggest to
follow this convention, even if this is currently no enforced on
JavaScript code.

Change-Id: I5c174ac7ea817c2252d6a054e7d5d2d662aa34ba