Do not double quote namespace numbers in the URL
The URL currently contains "namespaces":["1","2","3"], which is
(in the worst case) encoded as:
%22namespaces%22%3A%5B%221%22%2C%222%22%2C%223%22%5D

Reducing the number of (not strictly needed) special characters
results in "namespaces":[1,2,3], which encodes to:
%22namespaces%22%3A%5B1%2C2%2C3%5D

This is pretty significant (6 characters less per namespace).

I'm intentionally not changing the internal representation, the
contrary. This patch makes sure the internal representation is
always an array of strings. This was not guaranteed before.

Bug: T188509
Change-Id: Icbc6cc4b6c46e2c9327a64ddecfc55a91da79e83