Add missing test cases for GetPreferences hook handler
The hook handler was introduced late via T210479, but no test was
ever written for the new code paths.

In this patch I'm also removing the test case for the
ResourceLoaderTestModules hook handler. This is the code (soon
obsolete) that makes the QUnit tests run. It's a little pointless
to test code that is only there to run tests. I added the test
case back then only to raise the test coverage. Later I figured
it's better to just mark this QUnit bootstrap code with
@codeCoverageIgnore.

Bug: T210479
Change-Id: I24e161bd2e4c20bc3fec6a7308956a7d35a53093