Unify code requesting the main config from two sources
SpecialPage::getConfig() is documented to be a shortcut to fetch the
main configuration. It's identical to MediaWikiServices::getMainConfig().
Now note that the code was previously doing both. I found this confusing.
If it's meant to be the same config, it should look like it is using the
same config object in the code.

Change-Id: I5a8cfc4b5961acad35954ea049a25c00937f4155