Respect true/false settings in namespaces to be searched
This code never checked if the namespaces that were configured were
enabled or not. Both the user options and the global config use
true/false to indicate if the namespace should be included.

There were multiple tests asserting the broken behaviour, might
need further review to determine what was supposed to happen here.
This patch makes AdvancedSearch treat namespaces to be searched
the same as mw core.

Bug: T219455
Change-Id: I80b67a7747368b78950ea4a1b14bdaecec163836