Fix TitleCache failing for strings that contain dots
For example, the two strings "a.b" and "a.c" resulted in the same cache
key. The reason for this misbehavior was the getName() method that returns
a file name without extension, and without namespace.

Same for strings that start with a known namespace. This was stripped and
not considered in the cache key, also resulting in bad cache hits.

Change-Id: I3d7f783189ef514ad109ebaaf8ae17e725cf7c10