Avoid initial 302 redirect when no search is done anyway
This is relevant when I click the magnifying icon in the upper right
corner without entering anything. The search= parameter is empty then,
and no search is executed. Which namespaces appear in the URL is not
relevant in this situation. It won't change the search results for
different users with different default namespaces, because there is no
search and no search results.

This saves a redirect and makes the page respond faster.

Change-Id: Ic3ea680d574c7287d21b8eacec9980f74137289c