Simplify HooksTest setup with real User objects instead of mocks
The production code does not care if the user is anonymous or not. It
calls User::getOption() in both cases and expects it to either return the
user's preferences, or the wiki's default settings in case it's an anonymous
user.

`new User()` creates a user that is technically anonymous in both cases.
In the first case this user does not have preferences, which triggers the
code to return the wiki's defaults.

Bug: T223709
Change-Id: Ifda0bc76c65c858de0c1696dc5d43b5220dcc674