Fix browser tests due to recent changes on OOUI and the logout
It seems OOUI changed a little bit. Before, the click was done on the
little down-arrow on the right side of the file type selector. Now, the
<button> element that makes up the bulk of the file type selector is
clicked.

In addition to that, changes to the way the logout step works also made
it necessary to wait for the logout to be finished before the next test
be can executed successfully.

Change-Id: Ib711936e9195c748253a1bd7f138a968ba137307