Fix selectAll() helper unselecting selected namespaces
The previous code just assumed that the first element in the list is
selected, but nothing else is. It was blindly clicking all but the first
element. This started failing because the first *two* elements are
sometimes selected because of a leaking Wikibase test (see Ifda0bc7 for
the same issue in a PHPUnit test).

The new code only clicks elements that are not already selected.

Bug: T223709
Change-Id: I194ebb0066bad2874449a182601cac9b6e95d697