Update placeholder height to minimize jumping content
This was updated latest in If549515. It seems some styling (most
probably something in OOUI) changed slighty in the meantime and made the
lazy-loaded AdvancedSearch UI slightly bigger.

I tested this with Chrome and Firefox, and found both behave the same.
The page content jumps down 2px because the final UI measures 122px, not
only 120.

Change-Id: Iaa7bbe7e2eafe0d4d8f7eb5bbbb121a15a9833dc