Fix duplicate expandable pane id and aria control
Accessibility team from the company I work in found multiple
issues in MediaWiki, including duplicated ids in multiple elements
which may affect usability of wiki using programs like narrator.

Advanced search includes a method that hard codes the id in an
element, and this method is called twice in init.

This update changes the ids to be dynamic depending on a variable
passed in the params.

Bug: T244295
Change-Id: I7eb13c8c0c1effdb39b2fdd5eda856d982ade996