Fix "[object Object]" bug in page name lookups
It took me quite a while to find this, but it's surprisingly easy
to explain now: It looks like the OOUI LabelElement mixin was
originally designed with plain text labels in mind. That's what the
{ 'label': … } config does, that's what calling setLabel() does,
and that's what the internal this.label property stores: plain text[1].

Later, support for more fancy HTML labels was added. A config
{ '$label': … } and a property this.$label was added, along with
setLabelContent() (replaces only parts of the this.$label element,
preserving the outer DOM element), and setLabelElement() (replaces
only the outer DOM element, but keeps the content).

Some of this was bleeding into the existing plain text label support,
but it's incomplete and confusing. For example, how is it useful to
store a jQuery element in this.label, when there is this.$label?

I believe this.label should never contain HTML, and all documentation
that says otherwise is wrong. Or: if this is an intentional feature,
the support is incomplete and broken.

TL;DR: LabelElement#setLabel() is broken. We can simply avoid the
broken parts by passing plain text and HTML labels separately.

[1] There is also support for functions that return a string.

Bug: T249167
Bug: T258337
Change-Id: Id3ef3025793b10c6d6f0dc796043bfacede6b4de