Fix lookup code failing on titles with invalid characters
I found this additional bug while working on Id3ef302. Steps to
reproduce: Try to enter the character "[" as a title in one of the
category or template fields. This triggers all kinds of exceptions
in the JavaScript console. It seems all code assumes that all user
input is a valid title. If it's not, mw.Title.newFromText() returns
null and new mw.Title() even raises an error. Nothing of this was
expected.

Bug: T249167
Bug: T258337
Change-Id: Ie2aa6cc5b8e45cb565bcc10a0db12fb4a886fc68