Remove empty …&advancedSearch-current={} from URLs
This reduces noise in all [[Special:Search]] URLs:
* When clicking "Search", the parameter is submitted as a
  hidden field and becomes visible in the browsers address
  bar.
* The pagination links at the bottom of the page contain the
  parameter as well.

By default, the parameter reads `advancedSearch-current={}`,
which means it does not contain any information. It could as
well not be submitted. This is what this patch does.

Note that the code that consumes this (in initState() in
ext.advancedSearch.init.js) is already fine with the parameter
being missing. There is no difference between submitting {},
submitting an empty string, or not submitting the parameter.

Existing links that still contain the parameter continue to
work as before.

Change-Id: Icf7f9b70104ffd7a12ab45ddfcf6d926bf8068ac