Make sure special sort methods don't "disappear"
Certain wikis support special sort methods that are currently
not hard-coded in ext.advancedSearch.getSortMethods.js. Since
I41781ae these are show with their raw string identifier in the
preview (when the panel is collapsed). But the
DropdownInputWidget still shows a mismatching "relevance".

This is only a visual glitch. Re-submitting the form does not
change the sort order to "relevance". Still it is confusing to
see "relevance". This patch works around this issue and adds any
unknown sort order to the end of the list.

Note this is merely a temporary workaround, but no full fix for
T236891. A sort method that is not one of the 3 pre-defined ones
still doesn't appear in the list when it was not already active
before. There is already a TODO in
ext.advancedSearch.getSortMethods.js that describes a possible
solution.

Bug: T236891
Bug: T239030
Change-Id: Ic3d6837edd9ea270e9bd6c69c86ac5070da4a574