Remove unused "greedy" option
This was introduced for the "prefix:…" keyword, but is unused
since 2017, see I043fc44 (T171843).

The problem with "prefix:…" is that it consumes the rest of
the query string (it is greedy). The "prefix:…" keyword must
go to the end of the query string because of this.

Instead, we use "subpageof:…" now.

Bug: T171843
Change-Id: I44d0d188597ee4aa4464e0f13a473e4858173e2e