Programmatically use sort method messages when they exist
The message key prefix "advancedsearch-sort-…" as well as
"advancedsearch-sort-preview-…" is unambiguous enough to be sure
we don't accidentially show an unrelated message. Instead of
hard-coding the message names, we can use the name of the sort
method to construct a message key. This is not uncommon, done
another time just a few lines below.

This allows to introduce more messages for existing, valid sort
methods without touching multiple places in the code.

This patch also adds missing messages for the "ascending"
counterparts of existing messages. These will already be used
when following a link that contains e.g. …&sort=last_edit_asc.

Bug: T197525
Bug: T239030
Change-Id: I24fd27b37c82b5fcf0cf2037f5c05013c2972a3a