Rewrite TooltipGenerator to be testable
I noticed there was never a coverage report generated for
this class at
https://doc.wikimedia.org/cover-extensions/AdvancedSearch/
This was because the test accidentially called the static
generateTooltips() method non-statically.

Instead of just fixing this call in the test, I decided to
rewrite the class to work with an injected MessageLocalizer
instead of the global wfMessage() function. This fixes
another problem: This test was the by far slowest in this
codebase (>1s!) because of the many messages.

Change-Id: Ie378ba708d1b731db3da06aded83674770f53d48