Inline small pieces of JS code for readability
This is split from patch I8d22e9f. The idea is to *improve* the
readability of this code by removing unnecessary intermediate
steps. The code becomes a bite more "functional".

A notable change is related to Array.sort() which sorts the
array in place, *and* returns it. Best practice is to not use
the return value.

Change-Id: I698f1f846c9b811e38f13e6f5e83e44509d3dc2d