Refactor wdio tests from sync to async mode
I needed to add a custom version of the utility class that has a
async/await compatible version of the waitForModuleState() method.

I also moved one namespace test that includes a user login to a
different spec to group these tests that need a login and impact
the pre-selected namespaces.

Bug: T300791
Change-Id: I90ce9bbd28f2b01908413ed42b9dc11c47574451