Merge and reduce selenium tests in searchpage spec
I moved some minor tests to the main workflow test and removed the
test that's checking if the popups on the info icons work and have
content. This felt like something that's not worth a browser test.

I also renamed the spec since it's now only about tests around the
namespace selection.

Bug: T338793
Change-Id: I32502c54d37ed27f7b0c561d3b232a52c3a9f3b9