Preview file type selection as extension instead of mime type
This is really only a bugfix. The file types can be selected from a
list that says "svg", "jpg", and so on. But the preview shows the
internal, technical value "image/…". That's really only because of a
bit of missing code and was not an intentional decision.

The longer "image/…" is also cut off in the preview pill in many
situations. While this issue might not disappear entirely it happens
less often when we strip the prefix "image/".

Bug: T179589
Change-Id: Ib130874a0851277143851bdf4e3be53fe9830fb6