Spell out whole category and template for suggestion box test
This is not an issue on the CI it seems, but it is on the beta
cluster and I have a simillar behavior locally. It's still testing
if the box appears and the category is found. Only the prefix search
seems to be not working correctly.

Maybe there's an indexing step missing. It's weird. I'm fine with
the pragmatic compromise here though.

Bug: T282219
Change-Id: Ic569162e53c16de19bf2bef2341b00b5bcbe9483