Remove non-standard border from top of the namespace dropdown
The purpose of the thick 4px border is to separate one pair of
namespaces from the next pair. None of the mocks in T202505 as well
as T202370 show a thick border on the very top. Both tickets refer
to the filters on Special:RecentChanges. There is also a namespace
selection with pairs of namespaces separated by a thick border, but
no thick border on the very top.

It really looks like this was just a mistake.

Bug: T202370
Bug: T202505
Change-Id: Ib89db4d5e560ea47ed8bc4830b9279d4a987a526