Add ARIA labels and descriptions for the dimension fields
The FieldLayout around these input elements is a custom build and
the default setup on where ARIA labels and descriptions are applied
to does not work here.

I thought it makes sense to apply the label on both input widgets.
The difference is hopefully apperent, because the dropdown reads
the values that can be choosen there and the following is just the
text input.

I put the description only on the operator input because from the
order it's tabbed through it should be sufficient to have it once,
espacially since it describes both fields.

Bug: T339908
Change-Id: I06d18b37ad9ef0ed7302f0e2652e9f7cbcf0549d