Convert search preview tags into a list with a description
Screen readers read the whole content of the expandable button. So
the included list of preview tags for the selected search options
is read without context.

To give screen reader users a better idea on what's actually shown
here, I transformed it into a list and added a description so it's
hopefully more clear.

Bug: T342245
Change-Id: Id41b30096f56f3ceb58a154270a19c6fb6963372