Add seperator to search preview tags
One annoyance noticeable when using a screen reader is, that there's
no distinction between label and value of the search terms
previewed.

I first thought about using text formatting but that's also quite
intransparent for a screen reader. So I ended up just using a colon
and it defenitly helps.

Bug: T342245
Change-Id: If1ffc4b200706b0ac415cedb0e8b62b9584652bb