Drop pointless greater/smaller-than symbol messages
The two messages are marked as "do not translate". Which means what
actually happens is that we map the symbol ">" to ">" and "<" to "<".
In all languages. That's a no-op.

The messages have been introduced via I47c00c3 in 2017. As far as
I can tell the idea was to give the translatewiki community the
chance to localize the symbols in certain languages. Looks like the
decision was made to not do this.

https://translatewiki.net/wiki/MediaWiki:Advancedsearch-filesize-greater-than-symbol/qqq

Let's get rid of the unused complexity.

Bug: T342245
Change-Id: Ic954b3b57c84b691f092fec274183aa9ac9b06e7