Remove untranslated equal sign message from SearchPreview
This is a direct follow up to Ic954b3b. Same argument. The
translatewiki community decided to never translate this and marked
it as "do not translate". It's effectively dead code because of
this.

For reference: The message shows up in 2 places: As a label on the
preview pill as well as a tooltip when hovering the pill.

Bug: T342245
Change-Id: Id481930f2b83541330bcfc3e2c7269ca1ef96bee