Remove loading spinner instead of just hiding it
Why keep the element around when there is no code that can ever bring
it back?

Also remove a whole bunch of dead CSS related to this. The original
idea was to use the #mw-searchoptions element as a visible (!) but
blank placeholder. This doesn't happen any more since height was set
to 0. This job is now done by the spinner element.

Change-Id: I43ff940b52246ea2baca7c377cd2eea07c024d7c