Don't render default sort order in SearchPreview
The preview pill feature was added in 2017 via T165295 and T171756.
But there was no sort order back then. That was added two years later
via T197525. It looks like the behavior for the sort order preview
pill was never specified and never fully tested. It had many
problems, see esp. T267842. We fixed most of them, but some remain:

* The user can still click the pill, but what happens (the panel
  expands) has nothing to do with the sort order.
* It's currently disabled (gray). That makes it look like the user
  has no way to interact with it. Which is true and false the same
  time (it can be changed, but it's impossibly to not have a sort
  order).
* Why is this presented so prominently in contrast to other defaults
  that are not?
* Of what use is the information? The pretty much only effect is
  advertisement for the fact that the sort order can be changed. Is
  this worth it?
* It's more irrelevant stuff for screenreaders to skip.

Bug: T347382
Change-Id: I790b45629d3f0d4a1af5dc32eaf4dc6b97ea551a