Hide empty namespaces in AdvancedSearch namespace dropdown
Why bother searching a namespace when we know it's empty? It's just
confusing for the user when we show them a long list, but some
entries never do anything.

An example for such a namespace is
https://de.wikipedia.org/wiki/Special:AllPages?namespace=711

There is a fine balance here as all possible perspectives can be
confusing:
a) At the moment the user can pick the namespace but then wonder why
   it doesn't work. Nothing is found, no matter what they try.
b) Hiding the namespace might be confusing for a user that knows the
   namespace exists.
c) An alternative might be to show the namespace, but disable it.

I argue that b) and c) are extremely unlikely. Why should users be
aware of namespaces that don't contain anything anyway? Even if a
(technical) user knows about such (usually technical) namespaces,
we can expect them to be able to figure it out, e.g. via
Special:AllPages.

Bug: T184850
Change-Id: I77d9afcae606d852e55788a5fd539477a94b7622