Workaround to center search terms label
The close box is throwing off the alignment of other elements, but
it's proving very hard to address that issue directly since its
layout is buried in oo.ui and includes many nested nodes with fragile
calculated sizes and positions.

This workaround causes the label for search terms (such as "These
words:") to be aligned with the other text, rather than appearing at
the bottom of the pill.

Bug: T252346
Change-Id: Id5d110ffe467f8388bf5373caa940a8b28cc363a