Revert "Revert "Workaround to center search terms label""
This reverts commit da1ef99b8c45aef3b9cd4c3f2edf59dad30c01f9.

Reason for revert: This is the 2nd attempt to merge

Id5d110ffe467f8388bf5373caa940a8b28cc363a

The original cherry pick got reverted, because we merged out
side of a backport window. Back then we thought that might be
possible, because the train was not executed yet. Our bad.

Let's do it the proper way now

Bug: T252346
Change-Id: If2bf487add608d8d0874b38dda0e03c4de4be6dd