Fix inverted "is namespace (not) selected" in the code
Not sure how this happened. It's like this ever since this code was
added in 2018, see If279aa9. It's really only a naming issue. I
could have renamed the method to "not selected", but I feel like
this makes the code unnecessarily hard to read. Instead I stick to
the name and invert the return value.

Also reuse the method where possible.

This is technically unrelated to T354107, just something I noticed
during investigation.

Bug: T354107
Change-Id: Ie63b17d5da48222c76f47c11417abb828c711b03