Drop never used abstract methods from SearchField class
This code was added as part of Ifa08503 between patchset 1 and 2,
see https://gerrit.wikimedia.org/r/c/496174/1..2. It looks like it
was an unused draft that was never used.

The next patchset https://gerrit.wikimedia.org/r/c/496174/2..3
mentions a createSearchFieldFromObject method that was meant to
make use of the abstract methods, but apparently never did. Instead
the elements obj.init, obj.layout, and obj.formatter that do what
the abstract methods are meant to do are passed as arguments to
createSearchFieldFromObject. Documentation for this was added via
Ied373dc.

There is a second method with the name "createWidget", but it's
unrelated:
https://codesearch.wmcloud.org/search/?q=%28create%28Widget%7CLayout%29%7CformatSearchValue%29&repos=Extension%3AAdvancedSearch

Change-Id: Ieaa96b3760af323150ca2b2a5f37a012f8651e21