Replace confusing "lookupId" with numeric namespace id
I always found it confusing to find the concept of a "lookupId" in
the code. What does it mean? Turns out this is the canonical English
namespace name. But it's lowercase, so it can be used as a key to look
up (and that's probably where the name comes from) the actual namespace
id via mw.config.get( 'wgNamespaceIds' )[ this.lookupId ].

This is somewhat convoluted, I feel. Why not just use the namespace id?

Three things are happening in this patch to fully resolve this issue.
1. Turn the helper function getTitle into a method so it does have
   access to the class properties. Note: Can't be named "getTitle"
   because that method name is reserved by OOUI's TitledElement mixin.
2. Replace lookupId with namespaceId, as described above.

Change-Id: I851452d4291125f4e48a599d39ac60085546141b