Replace obscure "self" function parameter in MultiselectLookup
It's unclear why it was done this way when this code was introduced
in I17a3016. We can pass only the "queryCache" and it will be a
reference to the original object, no problem. The only exception to
this rule are arrays in PHP. Maybe that's what cause the confusion
here? But the queryCache isn't even an array.

This patch also cleans up related code. Remove "self" in places
where it's not needed any more. We changed the code to not use
functions any more, but the => operator. A core benefit of that is
that "this." can be used in a much more predictable way.

This is split from I5b86ecf where I try to fix a much deeper problem
with this code.

Change-Id: I15b67fc787a77b2333b4cdaf85ab0926b6d98c63