Remove duplicate code from MultiselectLookup subclasses
The MultiselectLookup base class already does this. This duplication
is meaningless and confusing. The code was left behind when the base
class was introduced via I17a3016. The two subclasses already existed
at this point.

We could even remove the two subclasses and turn the very few
remaining differences into configuration. However, I would like to
do this after I851452d and keep the two subclasses for the moment.

Change-Id: If546481ed0dcbbe739df58e9c8cbeef4f92e496c