Use waitForExist correctly
According to https://webdriver.io/docs/api/element/waitForExist/

"Returns true if the selector matches at least one element that
exists in the DOM …"

Note the "at least one". The [ 0 ] in the previous code doesn't make
much sense with this in mind, I believe. It was added quite a while
ago via I82689b4.

Let's see if this change helps.

Bug: T338793
Bug: T384920
Depends-On: I6a31b3e2c58ca15e6d51e49be94b33759b972782
Change-Id: I3ba76bf630d64e4fbaaa3a9289ba70a121595c3e